# src/openapi.yaml
openapi: 3.0.0
info:
  title: TUI API
  description: API for controlling TUI applications
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
paths:
  /api/screen:
    get:
      summary: Get current screen content
      description: Returns the current terminal screen as an array of strings (lines).
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/input:
    post:
      summary: Send raw input to the terminal
      description: Sends text or key sequences to the running TUI process. Useful for typing commands or sending raw escape codes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: "The raw input string to send. Example: 'ls -la\\r\\n'"
              required:
                - data
      responses:
        '200':
          description: Input sent successfully
  /api/key:
    post:
      summary: Send a special key
      description: Request to press a special key like ArrowUp, Enter, Tab.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  description: "The name of the key. Supported: up, down, left, right, enter, tab, backspace, escape, home, end, pageup, pagedown, delete, insert."
              required:
                - key
      responses:
        '200':
          description: Key sent successfully
  /api/resize:
    post:
      summary: Resize the terminal
      description: Resizes the terminal dimensions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cols:
                  type: number
                rows:
                  type: number
              required:
                - cols
                - rows
      responses:
        '200':
          description: Resized successfully